+++ kelvin-api/tests/test_route_school_class.py
@@ -143,19 +143,24 @@ async def test_get(
 ):
     ou = await create_ou_using_python()
     sc1_dn, sc1_attr = await new_school_class_using_lib(ou)
+    mail_address = f"{sc1_attr['name']}@example.org"
     async with UDM(**udm_kwargs) as udm:
         lib_obj: SchoolClass = await SchoolClass.from_dn(sc1_dn, ou, udm)
+        lib_obj.udm_properties["mailAddress"] = mail_address
+        await lib_obj.modify(udm)
     assert sc1_dn == lib_obj.dn
     url = f"{url_fragment}/classes/{ou}/{sc1_attr['name']}"
     response = retry_http_502(requests.get, url, headers=auth_header)
     json_resp = response.json()
     assert response.status_code == 200
     assert all(
-        attr in json_resp for attr in ("description", "dn", "name", "ucsschool_roles", "url", "users")
+        attr in json_resp
+        for attr in ("description", "dn", "name", "ucsschool_roles", "url", "users", "udm_properties")
     )
     api_obj = SchoolClassModel(**json_resp)
     await compare_lib_api_obj(lib_obj, api_obj, url_fragment)
     compare_ldap_json_obj(json_resp["dn"], json_resp, url_fragment)
+    assert api_obj.udm_properties["mailAddress"] == mail_address
 
 
 @pytest.mark.asyncio
@@ -169,11 +174,13 @@ async def test_create(
 ):
     school = await create_ou_using_python()
     lib_obj: SchoolClass = new_school_class_using_lib_obj(school)
+    name = lib_obj.name[len(lib_obj.school) + 1 :]  # noqa: E203
     attrs = {
-        "name": lib_obj.name[len(lib_obj.school) + 1 :],  # noqa: E203
+        "name": name,
         "school": f"{url_fragment}/schools/{lib_obj.school}",
         "description": lib_obj.description,
         "users": lib_obj.users,
+        "udm_properties": {"mailAddress": f"{name}@example.org"},
     }
     async with UDM(**udm_kwargs) as udm:
         assert await lib_obj.exists(udm) is False
@@ -187,11 +194,44 @@ async def test_create(
         assert response.status_code == 201
         api_obj = SchoolClassModel(**json_resp)
         assert lib_obj.dn == api_obj.dn
+        assert api_obj.udm_properties["mailAddress"] == f"{name}@example.org"
         assert await SchoolClass(name=lib_obj.name, school=lib_obj.school).exists(udm) is True
         await compare_lib_api_obj(lib_obj, api_obj, url_fragment)
         compare_ldap_json_obj(json_resp["dn"], json_resp, url_fragment)
+        udm_obj = await udm.obj_by_dn(json_resp["dn"])
+        assert udm_obj.props["mailAddress"] == f"{name}@example.org"
         await SchoolClass(name=lib_obj.name, school=lib_obj.school).remove(udm)
         assert await SchoolClass(name=lib_obj.name, school=lib_obj.school).exists(udm) is False
+
+
+@pytest.mark.asyncio
+async def test_create_unmapped_udm_prop(
+    create_ou_using_python,
+    new_school_class_using_lib_obj,
+    url_fragment,
+    udm_kwargs,
+    retry_http_502,
+    auth_header,
+):
+    school = await create_ou_using_python()
+    lib_obj: SchoolClass = new_school_class_using_lib_obj(school)
+    name = lib_obj.name[len(lib_obj.school) + 1 :]  # noqa: E203
+    attrs = {
+        "name": name,
+        "school": f"{url_fragment}/schools/{lib_obj.school}",
+        "description": lib_obj.description,
+        "users": lib_obj.users,
+        "udm_properties": {"unmapped_prop": "some value"},
+    }
+    async with UDM(**udm_kwargs) as udm:
+        assert await lib_obj.exists(udm) is False
+        response = retry_http_502(
+            requests.post,
+            f"{url_fragment}/classes/",
+            headers={"Content-Type": "application/json", **auth_header},
+            json=attrs,
+        )
+        assert response.status_code == 422, f"{response.__dict__!r}"
 
 
 async def change_operation(
@@ -225,6 +265,7 @@ async def change_operation(
         change_data = {
             "description": fake.text(max_nb_chars=50),
             "users": [f"{url_fragment}/users/{user.name}" for user in users],
+            "udm_properties": {"mailAddress": f"{sc1_attr['name']}@example.org"},
         }
         if operation == "put":
             change_data["name"] = sc1_attr["name"]
@@ -241,7 +282,10 @@ async def change_operation(
         assert response.status_code == 200
         # check response
         api_obj = SchoolClassModel(**json_resp)
+        udm_obj = await udm.obj_by_dn(api_obj.dn)
         assert api_obj.dn == sc1_dn
+        assert api_obj.udm_properties["mailAddress"] == f"{sc1_attr['name']}@example.org"
+        assert udm_obj.props["mailAddress"] == f"{sc1_attr['name']}@example.org"
         assert api_obj.description == change_data["description"]
         assert {api_obj.unscheme_and_unquote(url) for url in api_obj.users} == set(change_data["users"])
         usernames_in_response = {url2username(url) for url in api_obj.users}
@@ -273,6 +317,37 @@ async def test_put(
         operation="put",
         school=ou,
     )
+
+
+@pytest.mark.parametrize("operation", ["put", "patch"])
+@pytest.mark.asyncio
+async def test_modify_unmapped_udm_prop(
+    new_school_class_using_lib,
+    create_ou_using_python,
+    url_fragment,
+    retry_http_502,
+    auth_header,
+    operation,
+):
+    school = await create_ou_using_python()
+    sc1_dn, sc1_attr = await new_school_class_using_lib(school, users=[])
+    change_data = {
+        "description": fake.text(max_nb_chars=50),
+        "users": [],
+        "udm_properties": {"unmapped_prop": "some value"},
+    }
+    if operation == "put":
+        change_data["name"] = sc1_attr["name"]
+        change_data["school"] = f"{url_fragment}/schools/{school}"
+    url = f"{url_fragment}/classes/{school}/{sc1_attr['name']}"
+    requests_method = getattr(requests, operation)
+    response = retry_http_502(
+        requests_method,
+        url,
+        headers=auth_header,
+        json=change_data,
+    )
+    assert response.status_code == 422, response.__dict__
 
 
 @pytest.mark.asyncio